+++ dep/recastnavigation/Detour/Source/DetourAlloc.cpp
@@ -19,14 +19,14 @@
 #include <stdlib.h>
 #include "DetourAlloc.h"
 
-static void *dtAllocDefault(size_t size, dtAllocHint)
+static void *dtAllocDefault(int size, dtAllocHint)
 {
-    return malloc(size);
+	return malloc(size);
 }
 
 static void dtFreeDefault(void *ptr)
 {
-    free(ptr);
+	free(ptr);
 }
 
 static dtAllocFunc* sAllocFunc = dtAllocDefault;
@@ -34,17 +34,17 @@ static dtFreeFunc* sFreeFunc = dtFreeDefault;
 
 void dtAllocSetCustom(dtAllocFunc *allocFunc, dtFreeFunc *freeFunc)
 {
-    sAllocFunc = allocFunc ? allocFunc : dtAllocDefault;
-    sFreeFunc = freeFunc ? freeFunc : dtFreeDefault;
+	sAllocFunc = allocFunc ? allocFunc : dtAllocDefault;
+	sFreeFunc = freeFunc ? freeFunc : dtFreeDefault;
 }
 
-void* dtAlloc(size_t size, dtAllocHint hint)
+void* dtAlloc(int size, dtAllocHint hint)
 {
-    return sAllocFunc(size, hint);
+	return sAllocFunc(size, hint);
 }
 
 void dtFree(void* ptr)
 {
-    if (ptr)
-        sFreeFunc(ptr);
+	if (ptr)
+		sFreeFunc(ptr);
 }
